Show unknown health when runtime status is unavailable
